Truncate a float vertex register, do not saturate it
The SST-1 latches a 12.4 fixed-point vertex. A float vertex register is
converted to a 28.4 integer and the LOW 16 BITS are kept.
`float_vertex_to_fixed` used `(f32 * 16.0) as i16`, and a Rust float-to-int
cast SATURATES.

The truncation is load-bearing. Glide and its games snap a screen coordinate
to the 4-bit subpixel grid by ADDING a magic constant of 2^19, which gives
the sum an ulp of exactly 1/16 and leaves the 12.4 value in the low mantissa
bits, then write the whole biased float. Tomb Raider Gold's 3dfx build writes
0x49001377 = 524599.4375 for x = 311.4375. Saturating returned 32767 for
every component, so every triangle collapsed to a point.

MEASURED on Tomb Raider Gold 3dfx at 8e9 cycles: 213,323 of 213,584
triangles rejected for zero signed area, every one of them with all three
vertices at the SAME point. After: zero rejected, 8,575,503 non-black colour
pixels stored, and the 3dfx splash renders with its textures. Descent II
3dfx, which ships the byte-identical glide2x.ovl, agrees: 214,060 triangles,
none rejected.

86Box's vid_voodoo.c takes the same low 16 bits through an int16 cast of its
float_to_int32 result.

The test writes the magic-snapped form the games actually use. It fails on
the old conversion.

+++ crates/izarravm-video/src/distira/raster_math.rs
@@ -189,8 +189,18 @@ fn sign_extend_24(raw: u32) -> i32 {
     }
 }
 
+/// The SST-1 latches a 12.4 fixed-point vertex. A float vertex register is
+/// converted to a 28.4 integer and the LOW 16 BITS are kept.
+///
+/// The truncation is load-bearing, not a detail. Glide and its games snap a
+/// coordinate to the subpixel grid by adding a magic constant of 2^19, which
+/// gives the sum an ulp of exactly 1/16 and leaves the 12.4 value in the low
+/// mantissa bits, then write the whole biased float. The register sees a value
+/// near 524288 and must return the low 16 bits of its 28.4 form. A conversion
+/// that SATURATES returns 32767 for every such vertex, which collapses every
+/// triangle to a point.
 pub(super) fn float_vertex_to_fixed(raw: u32) -> u32 {
-    ((f32::from_bits(raw) * 16.0) as i16 as u16).into()
+    (((f32::from_bits(raw) * 16.0) as i32 as i16) as u16).into()
 }
 
 pub(super) fn float_color_to_fixed(raw: u32) -> u32 {
